impl: have tasks and habits reset on a new day
Issue #42
Allowed for fake date to be added for testing purposes.  Moved functions to a separate file  for readability
